Shorter date formatting
To fit on the screen with large titles

+++ SurfTracker/Helper.swift
@@ -114,7 +114,7 @@ class Helper{
     // Take session time (Date) and return formatted string
     static func timeToString(sessionTime: Date) -> String{
         let dateFormatter = DateFormatter()
-        dateFormatter.dateStyle = DateFormatter.Style.long
+        dateFormatter.dateStyle = DateFormatter.Style.short
         dateFormatter.timeStyle = .short
         return dateFormatter.string(from: sessionTime)
     }
